Move newly tiled windows into place in the window's workspace
Usually the window workspace differs from the monitor workspace when
moving a window to a background workspace.
Previously onWindowCreatedTiling acted on the monitor active workspace
which caused windows to always appear at the end of the outermost node
on the target workspace. Now they appear relative to the last selected node.